Add license column to softwares, normalise footnotes
I was browsing the repo and upon reading the software files I wished it was specified whether a software was open or closed source, so I added a column that specified this alongside whether it is paid, in an "AlternativeTo" style.
As I was doing this, I found it prudent to add some clarification in the form of footnotes and having read the readme.md, I believed the current implementation of footnotes gathered on the one file was clunky markdown-wise (even if it was nice for the user) and didn't use github's markdown footnote syntax, so I normalised all of the footnotes using proper github markdown syntax at the cost of added redundancy. I also slightly reworded some footnotes to better suit their new context. I removed some that were redundant, namely the 1.17.x compatibility ones.
May have missed something but I'm pretty sure I got everything.